Utility: better test nested ConfigurationGroup assignments.
The original test didn't really verify that deeply nested
ConfigurationGroup instances get their configuration pointer properly
reassigned on copy or on addGroup(). Which they don't, leading to an
internal assertion on save.

Sigh, this whole thing needs a serious redesign with useless pointers
removed, but so far I didn't find time and I don't have an idea how to
do this in a backwards-compatible manner.