Utility: fix config pointer reassign in deep ConfigurationGroups.
It did so only for the first level, but not for deeper nesting. Both
failed tests pass now; interestingly enough such recursive configuration
pointer reset was done correctly in the Configuration itself, but not in
the base -- so that's now unified.

But effectively this pointer shouldn't need to be maintained at all,
it's just one extra moving part that will go out of sync and the only
purpose of it is to mark the configuration as changed. Instead the
configuration itself should go over the groups on save and check on its
own if a "modified" bit is set. 2007 me wasn't a good programmer it
seems.