CMake: properly trim trailing whitespace in corrade_add_resource().
The regex was attempting to do that, but of course it didn't work,
because the [^"]+ part right before was eating up the trailing
whitespace as well. And CMake doesn't seem to support +? (the non-greedy
variant of "one or more characters"), so I had to split the processing
into a dedicated whitespace tripping step, after which quotes get
removed and only then the filename gets used. That also makes it more
robust in case of strange corner cases like

    filename="Shut up", she said.txt

which before would silently eat the leading quote and fail to find the
file too.

And added a test as well, together with a filename that itself
contains spaces.