TestSuite: make it possible to have file/line info in instanced cases.
This way the output contains also the location from which the instance
data come from, and so one can click it in the terminal and go to where
the particular instance data was defined. Not just the command that
failed, but the data too.

Probably too obscure to be useful for anything except me, but
nevertheless. It saved me a lot of time, especially in that one 7000 LOC
test file with instanced test cases consisting of hundreds of items.