CMake: introduce CORRADE_TESTSUITE_TEST_TARGET.
This variable can be used to create a dedicated target for building just
the tests, meaning they're no longer built as part of the default ALL
target.

Which is especially useful on platforms where building tests takes AGES,
such as Emscripten. Before I was achieving this by disabling
CORRADE_BUILD_TESTS and then enabling it again, but -- who would have
thought! -- if it gets enabled again, it starts building all new targets
from scratch. Which isn't good. This is a much better solution.