package/archlinux: use a dedicated test target for Emscripten.
And only on Emscripten for now, because I still want to be able to type
just

    ninja

to have my whole native project rebuilt, including tests that usually
discover problems already at compile time. Not having that would mean
I'd defer the test build to the CIs or to package install, which is
usually too late.