Containers: add Size, Size*D, Stride and Stride*D typedefs.
Those used to be typedefs in the StridedArrayView class, which meant
having to type

    Containers::StridedArrayView3D<SomeRandomType>::Size

instead of just

    Containers::Size3D

which was *significantly* annoying. Given there will be a
StridedBitArrayView soon, reusing the exact same class, it makes sense
to have them directly at the namespace level. The original typedefs are
marked as deprecated and will be eventually removed.

Name-conflict-wise, I don't expect there being any Containers::Size
*class* that would conflict with those -- conversely, if there would be
any such class, the first use case coming to mind is exactly this, a
multi-dimensional size type.