Containers: use debug asserts in StridedArrayView::asContiguous().
There's actually no way around calling these in tight loops -- i.e.,
it's not possible to call it just once for a top-level strided view and
then iterate over items as ArrayViews. Which means that the users would
either have to suffer the extra overhead, which is bad, or rewrite this
API directly, which is *vastly* more error-prone than if it would have
assertions disabled for release builds.