Utility: use debug asserts in copy() of contiguous ArrayViews.
Those should have as little overhead as possible as they're often called
on tiny chunks of data.

OTOH, the StridedArrayView copy() and flipInPlace() still has the full
assertions. There the expectation is that they're called on large sets
of data and thus the overhead is rather minimal compared to the copy
operation itself. Those also rely on arrayCast(), which has full
assertions as well.