Containers: make it possible to create a custom StringIterable.
To serve certain use cases such as packed strings in Magnum's SceneData,
an additional context pointer is supplied to the callback. Which
inflates this type to a whooping 40 bytes, but as it's meant to be a
transient type, never really stored anywhere in a lot of copies, it
should be fine.