Containers: fix conversion of ArrayView to StringView.
Until now, conversion of e.g. an Array<char> to StringView was possible
only through explicitly calling the StringView constructor, with the
implicit conversion being disallowed because it required one more
conversion than allowed in an implicit conversion chain.

Simply copying the approach used in Iterable already, this also allowed
me to drop a workaround in the StringView(T*) constructor which would
get picked for an Array<T> instead.