Containers: preserve the Global bit in String::nullTerminated*View().
So a StringView converted back from such a String has
StringViewFlag::Global set as well. This makes it possible to avoid
unnecessary copies when a String is used as a storage but is then passed
further to other APIs as a StringView again, and those APIs need to
preserve the view lifetime.

It's also useful for testing that a StringView was correctly copied or
not when storing it.