TestSuite: stop using std::vector and std::pair in the internals.
Like, I don't even know why was that still there -- I was using
std::vector::insert() *only* because the API is shit and there's no
function for appending a list, so that definitely wasn't a reason to
wait until arbitrary insertion / deletion is implemented in growable
arrays.

Digging into this also made me realize that I'm unnecessarily making
quite a lot copies of the rather heavy TestCase struct, which I'm going
to fix next. But at least by using a Containers::Pair in a
Containers::Array, it can make use of realloc() for growing in-place
instead of copying everything to a new allocation every time the
capacity runs out.

Also, because Pair now properly warns when narrowing conversions happen,
I cleaned up the index type to be unsigned 32bit everywhere. With the
current layout it means there will be a 4-byte padding between this and
the TestCase struct, but that's going to get reworked in the next commit
anyway.