Utility: properly test limit integer values in TweakableParser.
Because -- yay -- switching to Containers::Pair from std::pair made the
compiler *finally* properly check the types, and warn when
potentially dangerous implicit conversions happen. Which it did, and
this test is a repro case for the bug.

HOWEVER, apparently just testing with full bit values is not enough, as
it will do the correct thing even in presence of the bug. So I'm testing
also with some other bit pattern, and I'm updating the JsonTest with the
same test cases because *what if* there would happen the same.