corrade-rc: don't use std::size_t in the --single variant.
Use just an unsigned int, same as in the general case. It doesn't need
a header to be included and compiling a C++ file that results in a
binary with over 4 GB is physically impossible nowadays.

This makes the generated files completely standalone, and with []{
changed to [] = { they could also be C-compatible, which isn't bad.