corrade-rc,Utility: implement null termination and alignment.
Both done in the same bulk of work as internally it's the same feature
-- just arbitrary padding between successive data.

Apologies for not having alignment working properly for overriden groups
-- it'd need to have dynamic allocator support (i.e., passing an
allocator to an arbitrary API) first, and hooking them up with Path,
which is a significant amount of work. For most cases it's fine already
like this, as the default alignment is 2*sizeof(void*) usually, and
people usually need just 4 or 8 anyway.