[wip] corrade-rc: allow shorthand filename specification.
If no alias, nullTerminated or align options are needed or the global
values are sufficient, having to write a [file] section for each file is
overly verbose. This allows to specify filename= options directly in the
global configuration section.

It's filename= and not file= in order to not have to further complicate
the CMake parsing routine that extract dependency information from the
file -- this way it doesn't even need to be changed.

TODO: while basic support for this was easy enough, the filesystem group
  override code needs to be adapted and tested for this, which involves
  copying the file somewhere temporary and changing the contents
  (because we can't make use of an alias in this scenario)
TODO: there's also a newly uncovered bug, see the TODO in Resource.cpp
TODO: this also needs to be documented