Cpu: fix compilation in C++20.
Took me a while to understand what is actually being ambiguous here.
Also the compiler told me that it'd attempt to use !operator== to
"emulate" operator!=, which makes me suspicious -- wouldn't this lead to
incorrect behavior when uncomparable values are being compared, such as
NaNs? I.e., for which both operator== and operator!= return false?