Containers: significantly rework BitArrayView and BitArray tests.
It doesn't make sense to test operator[] in constructor tests, it's a
complex piece of machinery on its own and has to be tested separately.
Which is what access() is for -- and there it should be tested
*rigorously*, not just some random bits -- otherwise it's too easy to
lose something important during a refactor later.

Besides that using a common bit pattern to test all APIs as well as
debug output; testing also the std::nullptr_t pointer + size overload.