Containers: prefer is_constructible for testing disallowed conversions.
With is_convertible it tests only implicit conversions, so if the type
would allow an explicit conversion, the test wouldn't catch it.

In some cases I still have to use is_convertible for some reason, not
sure why. Not a priority right now (it's all just tests), so I added a
TODO.