Containers: add deprecation TODOs for some StridedArrayView APIs.
As this container is something rather extraordinary and without
having any "grownup" version in standard libraries to steal ideas from,
its API grew quite organically. And some of the ideas I had in 2018 are
now redundant or possible to be done in a more straightforward, readable
and maintainable way.

But I don't know yet how to best handle backwards compatibility without
introducing even more code so this has to be a TODO until I figure out a
better way. New APIs, such as the upcoming StridedBitArrayView, won't
have the questionable functionality implemented at all to not have to
immediately deprecate it again.