Containers: new StridedBitArrayView container.
A bit-sized variant of StridedArrayView, nothing else. Common code
needed by both that was in StridedArrayView.h is now moved to
StridedDimensions.h.

Slicing and transformation API is pretty much equivalent to
StridedArrayView, except for slicing not allowing dimensionality changes
(which will be implemented differently via an "updimensioning"
constructor, or done by the user simply through operator[]) and negative
values in every() (which is doable using flipped() already, so there's
no reason to duplicate the functionality).