Utility: change Json's {parse,as}BoolArray() to {parse,as}BitArray().
Returning the new StridedBitArrayView instead of StridedArrayView<bool>.
I considered just changing the function to return a StridedBitArrayView
and keep the name, but didn't want to introduce breakages. After all,
the class is there for almost a year now, even though I still view it as
"new". Plus, depending on usability requirements, I might un-deprecate
the original if it proves to be still useful.

Oh, and I like the sliceBit() very much. Very convenient.