Containers: added "updimensioning" Strided*ArrayView constructors.
Because it's a simple and obvious operation (a 2D image is like a 3D
image with just one Z slice, etc.), and doing it manually way a pain and
a lot of useless assertions.

In case of StridedArrayView this is also expected to eventually replace
the dimension-changing capability of slice(). Which, however, adds new
dimensions at the end for some reason.