Utility: assert on source/destination size in copy() just once.
Instead of once in the 1D overload and then again in the 4D overload.
Together with the simplified updimensioning this makes the 1D/2D/3D
specializations pretty lightweight and thus they can be inlined in the
header, allowing further optimizations of this call.

I still might eventually reconsider changing the assert to a debug one,
but so far I don't think the perf gains are worth the lower safety. The
optimizations done in previous commits are just because copy() showed up
high in the profiler, and the reason was that it was called on *single
elements*, in a loop, which definitely isn't the way to go. When used
properly, the assert overhead is negligible.