Containers: take raw arrays instead of pointers for GCC 4.8 compat.
At first I thought I'd have to rework the previous commit to #ifdef
around GCC 4.8 and use Size/Stride for it everywhere but fortunately
this worked as well. Nevertheless, it needs an otherwise unnecessary
dimensions parameter this way, so it's still scheduled for cleanup once
GCC 4.8 is no longer needed.