Containers: store a void* inside *BitArrayView.
Originally I wanted to store a char* because it made it possible to have
constexpr slicing and bit access, and to not have to cast every time I
have to calculate a bit position. However it required to have extra
char* constructor overloads to make it possible to both construct the
class from an arbitrary data and from char* in a constexpr context, and
then there had to be yet another std::nullptr_t overload to disambiguate
the two. And ultimately it resulted in being unable to construct a
constexpr bit view on, say, a bool[], or std::uint64_t[], which is a
severe downside.

So I changed the internals to store a void*, and even with the casts the
internals are actually much simpler -- mostly due to not having to deal
with the extra constructor overloads. A downside is that slicing and
element access in a BitArrayView is no longer constexpr, but I don't see
that used much in practice (and StridedBitArrayView didn't have that
before either). For consistency data() returns a void* now as well,
which might be breaking for some users (sorry). Just cast that if
needed.