Containers: BitArrayView construction from a fixed-size array.
Plus a constructor that takes a fixed-size array, offset and size --
because I quickly realized that it's rather hard to construct the class
in various tests and examples and not run into some OOB issues. This
should make it a bit safer, hopefully.

And yes, it already uncovered a few range bugs in
StridedBitArrayViewTest. And some more places I changed to use the
shorter constructor variant now.