Containers: remove StridedBitArrayView pointer+offset+size constructor.
This one was originally added to mirror StridedArrayView, where it's
possible to create it with a pointer+size in the one-dimensional case,
to have a convenient way to create either a contiguous ArrayView or a
StridedArrayView with the same expression.

However, in the case of bit arrays there's also a bit offset in addition
to a pointer and size, which, together with the fixed-size-array
constructors introduced in the previous commit, makes this
convenience helper unfortunately clashing with constructors taking a
size and stride. And having offset+size accidentally interpreted as
size+stride is absolutely undesirable, so we give up on a
potential convenience in the name of safety.

And of course, funnily enough, certain assertion tests that were using
the old constructors kept compiling even after it got removed, because
they now relied on the array-to-BitArrayView or nullptr-to-BitArrayView
conversion instead, interpreted offset as size and size as stride, and
thus didn't trigger the assertions anymore.