Containers: don't compile the same in StringView and MutableStringView.
Instead, have the implementation in StringView only and delegate to it
from MutableStringView. Makes the debug build ~1kB smaller and has no
effect on the release build (because I assume the compiler/linker just
deduplicated it). But I think the main effect would be on compile times,
especially with bigger functions, this is mainly just a test that all
compilers are able to handle this.