Cpu: multiple CORRADE_ENABLE_* macros don't work on Clang after all.
I was testing it insufficiently -- while Clang before version 8 picks
the last only, version after 8 picks the first only, and because the
test was using CORRADE_ENABLE_AVX1 CORRADE_ENABLE_AVX, it inevitably
succeeded as it was the same as if just the first was specified.

The test is now extended to verify with *properly* independent macros,
POPCNT and BMI1, and while writing that test I discovered that this all
now works properly on GCC 12+, and not by an accident, as noted at

  https://github.com/gcc-mirror/gcc/commit/364539710f828851b9fac51c39033cd09aa620de

. So this compiler gets now promoted to a Good One, while Clang gets
extra bad reputation for randomly changing the behavior rather than
fixing it. The workaround is thus now applied for all Clangs (which
includes also clang-cl that understands the __attribute__() syntax too)
and GCC < 12.