Containers: implement BitArrayView::count() with a POPCNT+BMI1 variant.
The currently envisioned usage is for relatively small arrays, so the
basic 64-bit x86 popcount instruction is enough -- when the time comes
and this shows up in the profiler for 256+ bit long arrays, I'll look
into an AVX2 (and native AVX-512 POPCNT) variants. It's also not well
optimized yet, as the various TODOs show, but it's a good-enough start
to build upon further.

A WASM SIMD variant may come togehter with the AVX2 variant as WASM
SIMD has a vector popcount instruction, popcount is in the basic
instruction set and as far as I can tell it's being used for the scalar
code already (or, alternatively, it's not used even with
__builtin_popcountll()). A NEON variant will also arrive eventually,
once I discover the relevant extensions.

BitArray::count() delegates to this API, I deliberately didn't bother
with multidimensional and strided variants yet as I don't need them at
the moment. When implemented, those will most probably just try to
execute this variant on largest contiguous subranges, similarly to how
Utility::copy() optimizes strided copies.