Containers: add an ability to set/reset all BitArray[View] bits.
Again not on the strided variant yet. Interestingly enough, when
compared to memset(), zero-fill with memset() takes *a lot* more time
than setting all bits to 1 (which is then significantly faster than what
I have, for some reason). On GCC at least, Clang is ~same for both.

Not fully optimized yet, I have a TODO there for some bit tricks. This
might eventually be viable for optimization with the BMI1/BMI2
instructions again, but so far I don't think that's needed -- compared
to count() it shouldn't be that often used operation.