Containers: don't allow arrayAllocatorCast() on a nullptr Array.
Okay, GCC 12, this one was actually a helpful warning, thank you. If it
wouldn't be for all the other useless noise which made me ignore it, I
would even act upon it earlier.

The warning was saying

  array subscript -1 is outside array bounds of ‘char []’

for the doc snippet that was calling arrayAllocatorCast() on a
default-constructed Array, and then attempted to arrayAppend() into it.
I thought that was safe so the assertion had a special case allowing
nullptr arrays regardless of their deleter, but when I actually tried,
it crashed. So it's not allowed anymore, and the doc snippet is updated
to not imply that this would be fine to call on an empty array.