TestSuite: Compare::String with a primitive diff implementation.
The diff algorithm is just the most basic thing possible, as even that
is significantly better than having to compare stuff by eyeballing
terminal output, or worse, opening an external diff tool. Eventually I
may add collapsing of large common slices, or secondary diffing inside a
line, this is just the first step on unblocking my productivity.

It's implemented in a generic way so I can later reuse it for container
differences as well (and differences on characters inside a line). I
didn't spend any extra time on optimizing the algorithm further (apart
from not using a STL map and/or allocating a new one in every iteration)
as it's ran only on test failure which isn't really a critical path.