TestSuite: highlight also changed bytes in Compare::String.
Rather rudimentary at this point as it only handles the case of a
single-line difference, not if for example two consecutive lines are
both changed to something else. Maybe I'll figure that out later.

Also contains a rather nasty "handling" of UTF-8 characters, in which
case it just bails out of highlighting inside the line. Should be done
in a better way eventually.