Containers: use the popcount builtin on Emscripten.
Clang at -O3 recognizes the pattern and replaces it with the
corresponding WASM instruction, but doesn't do that on -O2 and lower,
and it's possible the pattern recognition might break eventually.
Thus, to avoid unexpected regressions, I'm using the builtin directly
on Emscripten.

On x86 this doesn't matter, as there's a native popcnt instruction code
path chosen if available. On ARM I still have to implement a dedicated
code path.