Containers: don't use a bool in Triple NoInit test.
It's just too fucking painful with GCC 12 optimizations. Like, I
regularly got an XPASS for this code:

    Triple<float, int, bool> aTrivial{35.0f, 3, true};
    new(&aTrivial) Triple<float, int, bool>{Corrade::NoInit};

    CORRADE_EXPECT_FAIL_IF(!aTrivial.third(), "...");
    CORRADE_COMPARE(aTrivial.third(), true);

HOW ON EARTH can the boolean value fail on check, saying it's false, but
then in another check be true?! And of course no amount of trying to
change it to !(aTrivial.third() == true), or CORRADE_VERIFY(), etc.,
fixed it. The boolean always read as two different values on the two
lines. Amazing compilers, or UB, or I don't know what.