PluginManager: better test for a plugin load failure.
Instead of an unresolved external symbol it nows references a symbol
from a library but doesn't link to it, which on Unix platforms cause a
failure on load.

On Windows the DLL has to be linked, but as long as it's not next to the
test executable and not loaded explicitly, loading the plugin will fail
as well. Which means this case is now properly tested on Windows as
well.