Containers: add an ability to create a custom Iterable.
Similarly to StringIterable -- I need to convert a
`std::vector<std::shared_ptr<T>> to an `Iterable<T>` in the python
bindings. Ugh. Don't ask.

Marking as experimental for now because I may need to add also the
context argument and/or explicitly pass stride and index like in case of
StringIterable.