package/ci: adapt to arbitrary unnecessary Codecov breakages.
They removed the Python uploader because "nobody used it" (it was in
the top 1% of pip packages), and the token-less upload stopped working
some weeks ago (and of course even an error is reported with a zero
return code, YOLO).

I thought I'd upgrade to the "new" uploader, but can't because it's a
fifty megabyte Node.js binary that doesn't even have an ARM version.
So, yeah, it's now using the long-deprecated bash uploader that is long
unsupported instead. Imagine that!

I hate all this technology. Everything is extremely janky, poorly
documented, and any "upgrade" is actually a massive downgrade that
makes everyone angry. What the fuck.