Utility: add BitAlgorithms.h with copyMasked().
The implementation is extremely silly and not really optimized more than
the naive approach at this point, but I needed something to avoid
duplicating the same bit of code every time I need to make a masked
copy.

Which is why it's marked as experimental. I'm also not 100% sure yet
whether this is a good naming, for example there could be a copyMasked()
that takes a mask for the destination and it would only differ by
argument order, or a variant that takes both masks. And then, depending
on how error prone it might end up to be, it might be better to name it
copyFromMasked(), copyToMasked() etc. Or maybe not. We'll see.