Containers: direct element access in multi-dimensional Strided*ArrayView.
Originally I didn't add those because I was hopeful that C++'s
*removal* of operator, inside [] would happen sooner. But no, it only
got *deprecated* in C++20, so it's still impossible to write

    a[3, 4, 5]

and one has to do things like

   a[{3, 4, 5}]

instead. It's great that some features added in C++11 are already
*removed* in C++20, but this one still not. I refuse to work around this
with operator() so this is how it has to be for now.

The StridedBitArrayView now thus also gets multi-dimensional set() and
reset() APIs for completeness.