CMake: allow dynamic builds on Android and Emscripten.
They still default to static because that's the least painful, but it's
no longer forced.

As a side effect, this also fixes the -fPIC option to be properly
enabled by default. Before it wasn't, leading to nasty linker errors
much later when linking the final *.so. This is probably a behavior that
changed with some recent NDK, because before it definitely worked,
somehow.