modules: disable -Wdangling-reference globally on GCC 13.
This is, sadly, the first completely useless warning on this compiler
that I have to globally disable. It used to be only MSVC what's coming
with exquisite suggestions like "we'd like you to know the behavior of
this compiler changed for this expression in 1997 to be according to the
standard" or "type is move only, default copy constructor was not
generated".

Now it's GCC as well, but that slowly started in version 11 already,
with new warnings that generate far too many false positives. This one
warns basically for all cases where functions returning T& are called on
temporary class instances. Might make sense in the C++03 days where it
was all a std::vector and libraries spent most of their time
recursively copying giant chunks of data for no reason, but definitely
not today, with *views*. A temporary view returning a reference isn't
something that should warn, that's the desired use case, and this
warning now gets printed for basically all uses of Utility::Json, among
other things. Not useful, so disabling globally.

https://gcc.gnu.org/bugzilla/show_bug.cgi?id=107532 claims it's fixed
for GCC 13 but I hope a better followup fix is made.