TestSuite: suppress bogus -Wmaybe-unused on GCC 13.
This is way too similar to strange bugs from the GCC 4 / MSVC 2015 days
where using a function argument in a {} constructor initializer instead
of () made the compiler think it's unused, or the compiler
forgetting the variable was used when a static_cast was applied on it.

Couldn't find any related bugreport in a cursory Google search. Not
worth my attention. Would probably warn in the other overloads as well,
will copy the workaround there on a case-by-case basis.