TestSuite: document why CORRADE_FAIL() doesn't add unreachable marks.
I was well past implementing and testing the whole thing only to
discover that TesterTest crashes with ... a double free() error (!!)
with the unreachable annotations in place. Turns out I forgot the
interactions with XFAIL macros, and the hint is *so strong* that it
makes the compiler give up on everything that could happen after such as
various RAII and other cleanups. Which is probably why it ends up
causing a memory corruption.

Can't omit returns after CORRADE_FAIL() then, sigh.