Containers: allow arrayAllocatorCast() on a nullptr array, but properly.
But differently. This is a better followup to a fix made in
5b4d36f892311b3c130fa0cffe77174d3d5536e6, which unfortunately made some
valid use cases assert.

The original behavior before that commit was clearly wrong because using
a growable deleter on a null array would crash as soon as the capacity
is queried. However, asserting that the array is non-null also isn't
right because there may be completely valid code that just happens to
end up with a null array with a default deleter even after calling
various growable APIs, and in that case arrayAllocatorCast() should
happen consistently with those and simply just propagate the
(default-constructed) array forward.

The test is now updated to show this use case, together with hopefully
a detailed-enough comment to ensure this valid use case doesn't get
broken again.