Utility: add JsonWriter::beginCompactArray().
And implement writeArray() using it. This makes it possible to drop all
redundancy in the writeArray() APIs and simply delegate them to write()
in a loop.

Use case for exposing this as a standalone function is being able to
write arrays in a compact way without a pre-existing array that could be
directly passed.