Utility: repro case for a particularly nasty String assertion.
This fortunately manifested only as an assert inside
Containers::String::nullTerminatedView() in release builds and not
something worse -- when value<Containers::StringView>() for a
non-existent value was called, it created the view over a temporary
empty std::string. Which is fine on its own because it's empty, the
problem is the presense of the NullTerminated flag in that case. As the
temporary empty std::string is SSO'd, the view points to a null byte on
the stack, and that null byte gets very likely overwritten in the next
few instructions, making the string no longer null terminated. Which can
have critical consequences.