Utility: fix calling find() or operator[] on an empty JsonView.
I added those two for convenience, but failed to consider that someone
might call them on empty arrays / objects. Worst about this is that if
it wouldn't crash, it would attempt to try to find given element in a
completely arbitrary parent token, leading to either a confusing
assertion or a silently wrong result.