package/archlinux: enable colored GCC output.
Ugh, what a mess. Imagine that flipping this flag triggers a rebuild of
the whole project. How are colored messages affecting the build output
in any way?!

Then there's Ninja, which is such a hilariously mismaintained project
that despite several feature requests it *still* doesn't properly pipe
through the compiler output which then makes the compiler disable the
colors. This all worked in makefiles for *ages*, so why it can't here?

Ideally, as a plan B, I'd just set some env var to opt in to color in
the PKGBUILD only, but of course there's no way whatsoever todo that.
There's the CMAKE_COLOR_DIAGNOSTICS env var, yes, but of course it
only affects CMake on the first ever run, so setting it to a different
value after will do nothing. The only way to achieve this is thus
enabling colors always, and then opting out if those are undesirable.

Hilariously enough, Ninja automagically strips away color output if it
detects pipe redirection, it's just that it's NOT DOCUMENTED AT ALL
ANYWHERE which made me suffer with uncolored output for several years
because I didn't know this was a feature.

There's actually a whole page dedicated to a strangely named
CLICOLOR_FORCE variable, made by the Ninja maintainer himself,
http://bixense.com/clicolors/, and that same page also begs people to
recognize NO_COLOR as well, yet -- guess what! -- calling Ninja with
NO_COLOR=1 will *not* strip the colors of the command being run. I just
can't. Ugh.