Utility: replace inline EM_ASM() with an external JS library.
This allows each function to explicitly specify its dependencies, making
it no longer depending on what a particular Emscripten version decides
to include by default, or forcing users to painstakingly fill the
EXPORTED_FUNCTIONS array when linking the final executable.

It also allows the code to eventually get conditionally included or not
with preprocessor branches, for example to not include environment
queries for code that won't ever access Node.js console.

What's rather stupid is the inability to conditionally include code in
the JS file with the preprocessor. I mean, there's a PREPROCESSOR run on
each JS library file, and yet such a basic feature is impossible because
the implementation is too dumb for anything except basic #if conditions.