Use Utility::move() instead of std::move() almost everywhere.
Except areas that are still full of nasty STL code, like Arguments,
Configuration or Utility::String. For those the cleanup will be done
together with moving away from STL.

And not using it in documentation or documentation snippets yet either,
to avoid confusion. That may eventually change, but not yet.