Invert the CORRADE_STD_IS_TRIVIALLY_TRAITS_SUPPORTED macro.
There's CORRADE_NO_STD_IS_TRIVIALLY_TRAITS now that's defined when the
previous macro wasn't, as that leads to non-silent errors when code
using it forgets to include the TypeTraits header. Which can be seen
by all the added missing includes right in this very commit.

The old macro is still defined when appropriate if
CORRADE_BUILD_DEPRECATED is enabled.