Containers: disallow Pointer conversion to non-trivial/non-virtual base.
Until now, if a Pointer<Base> was created from a Pointer<Derived> where
the Derived type had a non-trivial destructor, or if emplace<Derived>()
was called on Pointer<Base>, deleting such a pointer led to the Derived
destructor not being called at all. Worse, if the Derived type didn't
have any heap allocation, such leak wasn't even discovered by
LeakSanitizer.

Now such conversion or emplace() is allowed only if either the Base has
a virtual destructor (in order to call Derived destructor as
appropriate), or if the Derived type is trivially destructible (in which
case it doesn't matter). That's another potential breakage for existing
code, like in previous commit, so if it triggers for your use case, your
code was silently wrong in the first place.