Containers: rework the StaticArray DefaultInit constructor.
Made it use std::is_trivially_constructible for consistency with all
other cases that branch on trivial constructibility, which required
moving the CORRADE_NO_STD_IS_TRIVIALLY_TRAITS directly to configure.h to
avoid including TypeTraits done in a previous commit and also putting
the IsTriviallyConstructibleOnOldGcc to a header that's included from
both StaticArray.h and GrowableArray.h.

Additionally it no longer delegates to an ugly overloaded private
helper, but instead it uses SFINAE in a proper way, which I finally
managed to learn 12 years after C++11 was out. The preprocessor makes it
look nastier and longer than before but in practice it should be
shorter and easier for the compiler to deal with.