Containers: add explicit conversion constructors to Pair and Triple.
Getting rid of remaining std::pair instances and realizing I need these.
As with any other type conversions (especially with Magnum math types)
and in contrast to std::pair I'm making the constructors explicit.

On the other hand, if the explicit conversion is made, it's warning-free,
even in presence of a narrowing conversion, as the explicitness of it
implies that such conversion is intended.