Document that <utility> is deliberately not used for std::declval().
Let's be reasonable. An alternative I considered was to implement it
myself and put it into Utility/Move.h, but the "declval" name is really
nasty, I didn't find anything better in the whole five minutes of
thinking about it, and Move.h didn't seem like a good candidate to put
that while TypeTraits.h OTOH felt like a header that nothing else needs
to include.

And then I searched for where std::declval is actually used, realized
it's basically every container header as the conversion from/to
external types relies on it, which then led me to question how <utility>
was not needed for any of them until now, and then I checked and even
Microsoft's STL has std::declval defined directly in <type_traits> so
why even bother.