Utility: port the Unicode namespace away from STL.
I'm not keeping any std::string overloads anymore either, it's all
Corrade strings now. Existing code still using the STL variants of
narrow() / widen() needs to update, for the rest the backwards
compatibility headers should be enough.

Since the return type of the utf32() API is changed in a way that
requires user code to update anyway, I took that as an opportunity to
return an Optional there as well, failing in case the text contains
invalid characters instead of inserting invalid codepoints. In
contrast the prevChar() and nextChar() don't return an Optional because
they're low-level to suffer from accidental misuse in practice (and
skipping invalid characters instead of failing as a whole might be a
valid use case).