Utility: add CORRADE_CONSTEXPR20.
As usual, way more pain than one would expect because there's not even a
single compiler that sets __cpp_constexpr to 202002 in all versions that
have __cplusplus set to 202002. So I had to manually go through all
affected versions to pick ones that actually support the feature I need,
and only then decide on the __cpp_constexpr value to check for.